Cleaned up code for MSVC 2008. This included modifying project settings, fixing up the nitpicky stdlib secure warnings (I didn't fix all of these, took the easy way out with #define _CRT_SECURE_NO_WARNINGS on some). Some obsolete projects removed from the solution.

+++ AlertRig/src/Tuning/Tuning.cpp
@@ -170,7 +170,7 @@ void init_pages()
 	{
 	case tt_contrast:
 		// Save the min contrast value for this type
-		m_tuned_param_current = m_iSavedContrast = m_tuned_param_vec[0];
+		m_tuned_param_current = m_iSavedContrast = (int)m_tuned_param_vec[0];
 		break;
 	case tt_spatial:
 		m_tuned_param_current = m_stim.sf = m_tuned_param_vec[0];
